block chain vs bitcoin
hello  viewers this is created by me after having a deep research it may have some 
abbreviations hope you will like it


Bitcoin first appeared in a 2008 white paper authored by a person, or persons using the pseudonym Satoshi Nakamoto. The white paper detailed an innovative peer to peer electronic cash system called Bitcoin that enabled online payments to be transferred directly, without an intermediary.

While the proposed Bitcoin payment system was exciting and innovative, it was the mechanics of how it worked that was truly revolutionary. Shortly after the white paper’s release, it became evident that the main technical innovation was not the digital currency itself but the technology that lay behind it, known today as blockchain.

Although commonly associated with Bitcoin, blockchain technology has many other applications. Bitcoin is merely the first and most well-known uses. In fact, Bitcoin is only one of about seven hundred applications that use the blockchain operating system today.
excited about block hood here it is
The blockchain is a type of distributed ledger or decentralised database that keeps records of digital transactions. Rather than having a central administrator like a traditional database, (think banks, governments & accountants), a distributed ledger has a network of replicated databases, synchronised via the internet and visible to anyone within the network. Blockchain networks can be private with restricted membership similar to an intranet, or public, like the Internet, accessible to any person in the world.

When a digital transaction is carried out, it is grouped together in a cryptographically protected block with other transactions that have occurred in the last 10 minutes and sent out to the entire network. Miners (members in the network with high levels of computing power) then compete to validate the transactions by solving complex coded problems. The first miner to solve the problems and validate the block receives a reward. (In the Bitcoin Blockchain network, for example, a miner would receive Bitcoins).

The validated block of transactions is then time stamped and added to a chain in a linear, chronological order. New blocks of validated transactions are linked to older blocks, making a chain of blocks that show every transaction made in the history of that blockchain.The entire chain is continually updated so that every ledger in the network is the same, giving each member the ability to prove who owns what at any given time.
well that was all hope it helped you....